Refactored Chat component
* Removed "room" prop
*  Moved Chat component from core to coin machine folder
*  Reused ActionsPageComment for comment input
* Reused ActionsPageFeedItem for comment list
