--- src/modules/core/components/Chat/ChatMessage.css
@@ -1,25 +1,0 @@
-.main {
-  display: flex;
-}
-
-.main + .main {
-  margin-top: 10px;
-}
-
-.avatar {
-  margin-top: 5px;
-}
-
-.commentMain {
-  padding-left: 10px;
-}
-
-.username {
-  margin-right: 5px;
-  font-weight: var(--weight-bold);
-}
-
-.timestamp {
-  font-size: var(--size-small);
-  color: var(--temp-grey-4);
-}
