--- src/modules/core/components/Chat/ChatMessage.css.d.ts
@@ -1,5 +1,0 @@
-export const main: string;
-export const avatar: string;
-export const commentMain: string;
-export const username: string;
-export const timestamp: string;
