+++ src/modules/dashboard/components/CoinMachine/Chat/Chat.css.d.ts
