+++ src/modules/dashboard/components/CoinMachine/Chat/index.ts
