+++ src/modules/dashboard/components/CoinMachine/index.ts
@@ -1,0 +1,1 @@
+export { default as Chat } from './Chat';
